add coverage constraint and fix Blocks generators
By looking at the code coverage reports, I noticed that some parts of the code were left
untested, which was weird because they should be part be when looking at the propery.
Yet, I noticed we use 'Int' (n <- arbitrary) as the number of blocks to generate. Int
is unsigned and therefore, doing 'take n' with n < 0 boils down to doing 'take 0'. As
a consequence, our generators was very often generating empty list of blocks, and, in
the remaining cases, didn't always generate sequences of blocks that would trigger
certain parts of the tickingFunction. That is now fixed, and coverage constraints has
been added to make sure of that