Remove 'EpochIndex' and 'SlotNumber' wrapper and make use of 'SlotId' instead when it made sense
This is because we _can't_ actually represent a SlotNumber at compile-time as we would like. At a given moment in time, this type has some given bounds.
Yet, it can change at any moment and considering it fixed would be a mistake. Plus, we can already encapsulate most of the slotting arithmetic within
the 'SlotId' type, so we don't _need_ the extra ones for the EpochIndex and SlotNumber and we can simply avoid manipulating them overall.