Review modules organization and folder architecture for Network Layer
There was a bit of confusion IMO about where the network layer sits
(understandable because there's no clear layout of what the architecture
should / could be anywhere ...).

Also, I took the opportunity to remove the 'Rust' part of the module
names since this is quite irrelevant. We are programming to an interface
(a.k.a the http-bridge API, and the implementation of that interface
shouldn't really matter).

Doing so, I've remove the unnecessary abstraction of the chain producer
Monad. We won't be instantiating multiple network layer at once, so for
now, this adds an extra layer of complexity that isn't needed. In
practice we can just fallback to a plain data type that we instantiate
using one service (e.g. the cardano-http-bridge) or another (e.g. the
Haskell shelley node).