replace usage of 'mtl' with 'transformers'
There's no strong feeling about using one or another. Transformers is
slightly simpler as it doesn't introduce any of the type-classes and
functional dependencies that 'mtl' adds. Though, we only use the power
of 'transformers' so far, and thus, for consistency at least, let's
stick with this one and only pull 'mtl' if we feel it becomes a need