replace 'Coin' with 'Word64' in the 'TargetRange' type
We don't need these various targets to be 'Coin', some may actually not
be; in the end, we will select several coins, so it's not impossible
that the sum of their value will overflow the max coin value. In the
rare case this would happen, we can simply fallback to splitting the
output into two change coins.