Adjust transaction size calculation.
Seems like we are _sometimes_ off by one byte. I say sometimes because, the node will sometimes
be fine with our calculation, and sometimes compute a size that is one byte bigger than what we
compute. This is rather weird and to be honest, we don't have time anymore to look deeper into
this. Going for the simple approach here