Adjust integration tests transaction scenarios to cover multi-outputs
Fees are actually not necessarily fixed. They depend on the ix of the input picked. Since we generate all fixture wallets
with an initial transaction with 1K outputs, ix can go from 0 to 999, and therefore be encoded on 1, 2 or 3 bytes. This
depends on which input are assigned to which wallet. As a consequence, it is easier (and more correct) to look at a range
of fees, in case inputs ix gets encoded on 1 or 3 bytes