Fix a previously undetected bug in the specification.
The `encode` function was mistakenly expected (by the specification) to produce
an invalid Bech32 string, when given an upper-case human readable part. If this
invalid string is passed to the `decode` function, it fails (unsurprisingly) to
decode.

This change updates the specification so that the `encode` function is
now expected to produce a valid Bech32 string.