Allow a known-to-fail counterexample within the Bech32 character omission test.
In the case where the tail of a valid Bech32 string is composed of one
or more consecutive 'q' characters followed by a single 'p' character,
omitting any or all of the 'q' characters will still result in a valid
Bech32 string.

We amend the character omission test to allow for this case, so that
when it occurs, it won't cause the test to fail.