Allow a known-to-fail counterexample within the Bech32 character insertion test.
In the case where the last character of a valid Bech32 string is the
character 'p', inserting any number of consecutive 'q' characters
immediately before the 'p' will still result in a valid Bech32 string.

We amend the character insertion test to allow for this case, so that
when it occurs, it won't cause the test to fail.