re-enable prop_sequential_put for tx history
Problem here was that our generator would generate inputs, outputs and
transaction ids independently. While this didn't matter much with the
MVar implementation, it does create weird inconsistency with the
relational model because indeed, we expect inputs associated with a
particular tx id to not differ when associated with the same tx id.

Since a transaction id is actually a hash over the content of the
transaction, changing the transaction should yield a different id.
So now, we generate "fake" txid by computing them from the actual
transaction content (as done in production, we simply don't hash
it in the test as this has little value here but cost quite some
computing power).