Call 'cleanDB' before each property run.
We use to create a fresh new DB layer before the run of each property
because this was very cheap. Now that we do create a db layer only once
before a given property, we need to make sure that the state of the DB
is at least cleaned up between each property.