Use plural name for map of private keys within `Mock` data type.
The `Mock` data type has the convention that each map has a plural name
based on the type of value held within the map. For consistency with
this convention, the map of wallet ids to private keys should also have
a plural name.

Similarly, we can use `pks` to represent a map, but `pk` to represent a
value within a map.