Make DB QSM property `prop_sequential` fail if one or more tags are not covered.
Module `Cardano.Wallet.DB.StateMachine` defined a set of `Tag` values to
represent various test scenarios that are considered important or
interesting. It also supplied a `tag` function to match `Tag` values to
arbitrary sequences of commands.

However, the test suite didn't actually check that all possible `Tag`
values were covered during execution. This created a risk that we might
actually not be covering all the cases we'd like.

This change makes the `prop_sequential` property fail if one or more
tags are not covered.