reduce usage of 'fixtureWallet' in the integration tests
There's a non-significant cost with the use of 'fixtureWallet' since it needs to restore a funded wallet
and wait until restoration. We also only have a limited number of those wallets, so it's better to keep
them for when we actually really need funded wallets. When testing parsing errors or wrong headers, there's
no need for demanding a wallet with funds