Parameterize wallet core over the transaction data-type
This commit defines a new abstraction layer via the class

    class (NFData (Tx t), Show (Tx t), Ord (Tx t)) => DefineTx t where
        type Tx t :: *
        txId :: Tx t -> Hash "Tx"
        inputs :: Tx t -> [TxIn]
        outputs :: Tx t -> [TxOut]

This allows for leaving the transaction shape rather 'open' and
define by whoever is implementing the wallet layer, so long as
it provides a few accessors on the transaction type and some
mandatory type-classes used across the wallet.

This allows for supporting both cardano-http-bridge and jormungandr
at the same time (and likely, add support for shelley-haskell
without too much issues).

One point remains fairly hard to abstract at the moment and it's the
DBLayer, or more exactly, the SQLite implementation. Since it's defined
in `cardano-wallet-core`, we can't -- in theory -- assume anything
about the shape of the transaction data-type. This makes it hard to
prepare the shape of the database without any prior knowledge of the
data-type. We cheat a bit by defining an extra class in the SQLite
module:

    class DefineTx t => PersistTx t where
        resolvedInputs :: Tx t -> [(W.TxIn, Maybe W.Coin)]
        mkTx :: [(W.TxIn, Maybe W.Coin)] -> [W.TxOut] -> Tx t

This assumes a specific shape for the transaction but is still flexible
enough that this assumption isn't too pervasive in the SQLite code. In
practice, this is _sufficient_ for our needs, so we may not bother any
further.