adapt Jormungandr LauncherSpec for integration tests scenarios
This is a first draft that spins up a Jormungandr node, and starts a wallet server to talk to it directly.
The setup is definitely simpler than for cardano-sl, and also much faster!

One thing I couldn't do was to generate many UTxOs per wallet. The genesis file in newer versions of
Jormungandr doesn't support having duplicate initial funds. A possible fix would be to generate use the
first 10 addresses of each wallet, instead of re-using the same address.