use first 10 addresses in faucet wallets (instead of 10x the first one)
This is because with Jörmungandr, we can't have an initial UTxO in that form: the input txid
in the genesis file are generated from the initial configuration address and value. So, having
duplicated 'fund' sections would lead to multiple transactions with a same id, which is
of course forbidden. Instead, we can generate and use the first 10 addresses of each fixture
wallet, but also have to do it for http-bridge