turn off stderr in merge builds in CI
This is an attempt at solving tests hanging in CI for no apparent reason.
What it weird is that, there's no clear sign of failure in CI, it's like
the test runner just stop outputting anything and freeze. This could be because
of too many things are actually logged in stdout, with a format that is 'too
interactive' for travis.