Protect against accidental construction of invalid arbitrary `UTCTime` values.
The constructor for `UTCTime` accepts an arbitrary offset from midnight, and
doesn't provide upper-bounds checking. If the offset is greater than one day
(allowing for leap seconds), it's possible to create an invalid time value.

We protect against this by constructing `UTCTime` values with the `addUTCTime`
function, which guarantees to provide correct handling in the event of overflow.