use 'fixtureWalletWith' in TRANS_LIST_RANGE tests
The fixture wallets for Jormungandr and http-bridge have actually rather different setup.
Although they have the same amount of UTxO, one is made of 10 txs, and the other of a single
tx with 10 outputs.

Yet, the test really expect wallets to have a single transaction, and therefore, it is
necessary to construct the wallet in such way