only store checkpoints for unstable blocks during restoration
We restore blocks by batches, but storing a checkpoint for every single block is very
expensive: storing a checkpoint is roughly 10 various db queries, so it's 200k requests
per epoch, each of them likely useless since we can't rollback that far!

Instead, we only store a checkpoint for the each last block in a batch or, for unstable
blocks.