replace pending transactions with discovered transactions
The idea of having to store two types of the same transaction, one
pending and one in ledger. Especially because we have to filter the
duplicates out by traversing I believe we can handle rollbacks nicely by
keeping a simple row. Here's the idea:

• When transactions are created, we mark them as pending and insert
   them in the DB, with a slot_id equals to the moment they got created (no
   change compared to now)

• When transactions are discovered, we replace / overwrite the previous
   transactions with the new ones, which update their state to "InLedger"
   and the slot_id at which the transaction was discovered. Incoming
   transactions are simply inserted.

• Now, in the case of rollback:

   • We discard every Incoming transaction beyond the point of
      rollback, they were not created by us and we haven't discovered them.

   • We update every Outgoing transaction beyond the point of rollback
      in such a way that their slot_id becomes the point of rollback, and
      their status become Pending again.

:information_source: This also covers the case of transactions going
from the wallet to itself, because they are marked as Outgoing; our
semantic for these directions is:

• Outgoing The wallet balance decreases.
• Incoming The wallet balance increases or stays the same.

Conceptually, it means that we interpret slot_id as the last moment the
transaction has changed status. There's no pruning needed, because
there's no duplicated rows. When we rollback, we effectively forget the
discovered transaction, and re-insert a pending transaction so the
approach is also quite intuitive IMO.