use nearest checkpoint for new tx meta slot + correctly keep track of pending history in db model
Keeping track of the tx history is kinda awkward, because it is part of the checkpoints but, stored alongside the actual transaction history!
Yet, what happens when we decide to insert a checkpoint that already exist but that has a different pending set?

-> Most reasonnable is to stick with the semantic of 'put' and replace the existing checkpoint with the one that is given. This implies cleaning
   up the old pending transactions before adding the one given by the new checkpoint.

Also, when we rollback, we do re-create some pending transactions and therefore, in the DB model at least, we need to make sure that all previous
checkpoints know about them