track pending transactions outside of the wallet model
We do store pending transactions next to wallet checkpoint, so there constantly
exists two sources of truth regarding pending transactions; this is in general
something we want to avoid in a data-model. Consequently, the database model
and some of the manipulation we have to do in order to accomodate for pending
txs feels awkward. Moving their management outside of the primitive models
simplifies pretty much all areas and remains valid.

In a next step, this will also help us get better testing coverage, especially
regarding the manipulation of pending transactions!