Move wallet engine and worker registry inside their own module
I've moved pretty much everything we currently call 'wallet layer' into a new module 'Engine' with the exception
of functionalities that aren't tight to a particular wallet. So, everything in the wallet 'Engine' expects a
context that contains a handle to a database, such a handle will be provided by the registry and registered
workers. This should simplify the manipulation and testing of isolated wallets, because from the engine perspective,
everythings happens as if there was only one wallet. Then, the 'Cardano.Wallet' module will now make the bridge
between the API and the isolated engine by pulling wallets from the registry, and running whatever needs to be
ran within the worker context (i.e. with a DB connection).