cli: Assume utf-8 output encoding
The cardano-wallet sources are utf-8 encoded, and so all of our
strings are utf-8 encoded.

If cardano-wallet is run under a locale which is not UTF-8, it will
crash as soon as it tries to print a UTF-8 character.

With this change, it won't crash, but it may emit mojibaked characters.