assert on the source wallet in TRANS_DELETE_01
The source wallet is a fixture wallet, so there are incoming transactions which may come from different sources.
Since we list transactions with no particular ordering, we can't really guarantee which one will be the first item.
Using the destination wallet instead makes it non ambiguous as there is only one transaction in this wallet