Use `read` instead of `rational` within `FromText Double` instance.
This matches the existing use of `show` within the `ToText Double` instance.

The implementations of `show` and `read` for `Double` are expected to obey the
following property:

```
read . show = id
```

The implementation of `Data.Text.Read.rational`, on the other hand, differs
slightly from that of `read @Double`. This difference occasionally introduces
small errors across roundtrip serialization & deserialization.

For example:

```
> a = -9.0e-3 :: Double
> a
-9.0e-3
> read @Double (show a)
-9.0e-3
> rational @Double (pack $ show a)
Right (-9.000000000000001e-3,"")
```