move Network, address conversions from public key back core
We currently use an opaque representation for `Address` which is very
convenient as it allows us to pass them around without really having to
worry about their internal structure.  Serialization and deserialization
is easy, and we can defer inspecting the internals only when really
needed.

Knowing the internal of addresses really matters only in a few places:

1. When performing random discovery, we need addresses to match the
Byron format (see
https://github.com/input-output-hk/cardano-wallet/wiki/About-Address-Format---Byron)
and carry their derivation path as a payload.

2. When performing sequential discovery, we need to be able to extract
the spending key (or a hash of it) from the address. This is possible
for both Byron and Jörmungandr addresses.

3. When generating change addresses, we need to know how to convert a
given spending key to an address.

4. When displaying a list of addresses from a list of known spending keys.

Our current codebase makes the distinction between `SeqKey` and `RndKey`
and implicitely assumes Byron addresses and random discovery for
`RndKey` and Jormungandr addresses with sequential discovery for
`SeqKey`. These assumptions are however false, since it is possible to
have sequential discovery also for Byron addresses (like Yoroi does).

In addition, the wallet engine is currently parameterized over a `s`,
`t` and `k :: Depth -> * -> *` type, which suppose then that a given
engine may only use one type of address per target. This again is a
false assumption since Jörmungandr comprehends both formats.

So, we have to review how we actually look at addresses and key
derivation such that:

- [x] Address format only depends on
  - The era / key type it's based upon
  - The network discriminant in use

- [x] Sequential discovery can be used with addresses or any format