implement first draft of binary decoder for legacy UTxO
The format for legacy funds is different from the format used for
standard genesis fund. The message payload contains only a list of
txout (address, value) with no associated inputs? It is still
unclear to me what these inputs should be and how we could realistically
handle these legacy funds