Remove now superfluous indirection for defining Tx structure
Undo the abstraction to introduce another one later that'd be more
satisfactory. Actually, with the Haskell nodes coming in, we'll likely
have more than just the transaction type that is going to be different.
So the little trick we did for the http-bridge vs jormungandr might not
sufficient. And, in order to introduce a proper abstraction here (and
probably avoid concrete types altogether in the core code) it's
necessary to first remove what is now a wrong abstraction

Co-authored-by: KtorZ <matthias.benkort@iohk.io>