distinguish first epoch handling when listing stake pools
This is because, during the first epoch, there's not enough
data to provide accurate measures for all stake pools. Some
pools may look "bad" at the beginning of an epoch while some
others may look good just because they were by chance, elected
for the first blocks.

To circumvent this, we simply return pools in an arbitrary order
during the first epoch, to make sure that results are not biased.