Add memoization to speed up release-lib.nix
`release-lib.nix` calls `packageSet` many times. This
results in lots of repeated work, both during eval stage
and when the jobs are run.

By memoizing the calls here we get a very significant
speedup of the eval time on hydra.

Long term, someone should look into fixing or replacing
`release-lib.nix` so that it does not call the `packageSet`
argument any more than is necessary.