remove usage of 'waitForNextEpoch' to check for the next epoch time
This function uses the NODE tip to check whether we have moved to the next epoch or not.
However, the next epoch start time is about where the network is at, so we are comparing
apple and oranges here. Instead, the test now waits for the time indicated by the API and
query the network tip right away to check whether we have indeed moved to the next epoch