make api handlers more polymorphic to reduce duplication
It was okay-ish to duplicate handlers for Byron random wallets in order
to quickly iterate and develop handler for those. However, a lot of the
logic was duplicated in the process and, with the upcoming support of
Yoroi wallets, we would need even more. However, this commit take
re-locate the discrepency between wallets types to arguments passed to
other more generic functions. This way, we can group the logic for
top-level functions and adjust how wallets are represented when needed.

This enables rather seamless support for Yoroi wallets, pretty much out
of the box.