Add a new key target 'Icarus' to capture the era in-between Byron and Shelley
Icarus keys are a variant of BIP-44, but working on top of Byron addresses. So they are an
hybrid form between Byron and Shelley. Note that we only support the 'Mainnet variation of them.

TODO:
Port relevant unit tests for testing Icarus key fully, in particular, golden tests with Yoroi.