Make creating and registering a worker into a single operation.
This change adjusts the worker registration lifecycle so that it's no
longer necessary to create and register a worker in two steps. Instead,
a single call to `register` will both create a worker and register it
within the specified registry.

In addition, we now provide automatic garbage collection of stale worker
references from the registry when worker threads terminate prematurely.