Make API path params capture tests as unit tests
These are needlessly taking time in the integration tests and above all, are duplicated all over the place
for all endpoints, whereas what we really care about it here is to test the behavior of the API when
capturing path params that are invalid. The same thing should be done for all the headers tests and basically,
any test that doesn't actually reaches the API handlers!