Fix threads synchronization in around all
The issue was that, the main thread was not waiting for the 'resource thread' to be done and to
have cleaned up the resource. In our integration setup, the resource actually spawns an
underlying process! So, not waiting for the resource to be cleared means leaving a process
running even after the tests have exited! So, I added another point of synchronization between
the threads to make sure that the main thread would only exits AFTER the resource has been released.