rework the checkPassphrase / encryptPassphrase phantom types
Wallet using a passphrase that is encrypted with the legacy scheme needs to go through some special
transformation before being encrypted. In order to present programmer mistakes here, we leverage a bit
the phantom types for 'Passphrase' such that passphrases obtained from the API are marked as 'raw' whereas,
every function from the 'AddressDerivation' module requires a passphrase to be marked as 'encryption'.
Such passphrases can be obtained via 'preparePassphrase' which does nothing but a simple type coercion,
for legacy passphrases it'll do some preliminary transformation.