move list databases to enable server calling it from API handlers
When a wallet worker dies, it is removed from the registry and therefore, is no longer
available for listing. Yet, the file might still exist on disk, so looking at the database
files is more reliable than looking at the registered workers. Should a worker be dead
however, operation on a wallet should be degraded