use Haskell magic and strong _family_ bonds to remove the need for passing discriminant as option
The main trick here was to make the 'Api' type poly-kinded and then allow some of the problematic types
to be defined as family of the given type parameter. This enables us to use more flexible data-types in
the Haskell typed client (like plain Aeson.Value) since the client is in the end simply returning JSON.

what we loose however by doing this is the ability to catch _early_ (i.e. in the command-line client code)
that an address may be of the wrong discriminant. This is deferred until the server sees the address and
reject the API call. I think this is really acceptable.