disable Mux & Chain sync traces
These traces, even if only printed at lower severities, are causing tremendous
slowness in the wallet and in particular in the restoration loop. They are
quite verbose, even for a 'Debug' output actually.

Also bumped the Handshake min severity from Debug to Info as this trace is
quite useful and only occuring once when a new client establishes a connection
with the node.