repair fingerprint conversion for Icarus addresses
This was forcing 'Mainnet by default because of how we consider Icarus addresses
on the ITN. However, now that we have integrated with cardano-node, this exception
is no longer valid.

Note that we can't really take anything else than the full address for a
fingerprint on Byron addresses because both the public key and the
network discriminant gets hashed in the address root! Thus there's no
way to recover them from a raw address. This also forces us to provide a
network discriminant in order to compute the fingerprint!

fixup