Make sure to only generate addresses using the Hardened index convention
It seems that the old / legacy wallet has trouble digesting addresses that are using indexes in the
soft convention. Note that the derivation is always hardened, regardless of the index value but
the convention was switched a while ago on -sl to use only indexes beyond 2^32...

Note that there are still many addresses on the current Mainnet that do not abide by this convention
and these addresses seems to break legacy wallet. So, if we can avoid generating potentially
problematic addresses in the new version, that's good :) ...

We still recognize address and account indexes across the whole domain though