disable flaky test built on race-condition
This scenario needs some rework as it'll likely keep failing "randomly"
if we happen to submit the transaction too late in the slot, leaving no
time for it to be forgotten.

This test seems rather _fine_ on Linux though, so it may be due to some
slowness on Windows and the delays induced by the command-line / process
spawning.