refactor: merge 'remainingFee' and 'rebalanceChangeOutput'
The former actually makes assumption based on the output of the later. Testing these
function in isolation makes little sense. Grouping them into one function makes it
easier to understand the surrounding context and what is happening.

Further refactoring is needed in order to now fix the issue with dangling fee