remove passphrase validations on endpoints using passphrase
It is still required to provide a valid and "strong" passphrase
when creating a wallet though. But if a wallet has been created
with a passphrase that is shorter than the expected validation,
then the API won't reject it and simply go forward trying to use it.

This only use-case and possible way for introducing a shorter
passphrase is by importing a legacy wallet from an already encrypted
private key. In this case, we can't control rules regarding the
passphrase and, the passphrase might as well be one single character.