Radically re-work CliKeyScheme & add bech32 encoding
This commit abandons the philosophy of having all logic in CliKeyScheme
and all encoding happening using mapKey.

Also: Support reading key inputs from stdin

Only con is that ommitting the key now no longer displays help-text
which could be confusing for users.

== Fixups ==

=== use T.encodeUtf8 / T.decodeUtf8

...rather than e.g. `B8.pack . T.unpack`.

Was told it was the proper way.

=== rename non-recursive `go` to `action`

If we were to only use stdin for key inputs, we could probably use
do-notation, and not have to use this "yet-another" helper.

=== parse simple `KeyEncoding` instead of function