Move 'StakePoolMetadata' inside '..Primitive.Types'
Although the underlying implementation of 'how we get metadata' is different, the overall presentation type
remains the same and is fortunately based on the formal specifications description in the case of both
jormungandr and cardano-node. This type can therefore live in the shared part (which is actually mandatory
since also used by the API layer).