nix: remove socket files from source trees before building
Any cardano-node.socket files present in the source tree will cause
an error in the nix build.

This happened in the nightly build. I do not know why the `git clean`
command that Buildkite ran did not remove cardano-node.socket from the
git checkout directory. But this change will prevent the error.