review RELEASE_TEMPLATE to constuct changelog from PR + labels
We used to organize the CHANGELOG by milestones, but we are now organizing it by 'type of change' (improvement, bug fix, addition). Therefore, the current thing called CHANGELOG is redundant with the other. I've tweaked the 'make_changelog' script to build a changelog according to these categories, provided that PR were labelled accordingly.