Remove migration tests out of release jobset
Since migration tests include several old revisions of cardano-wallet,
they are costly to include in the release jobset, in terms of eval
time and memory usage. Furthermore, if there are improvements to nix
evaluation speed in new revisions, they still do not apply to old
revisions.

So this moves the full sequence of migration tests out of the release
jobset. The migration test driver for the current revision is still
built (and cross-built). Therefore, builds of old revisions will be
available in the IOHK Hydra cache, where they can be downloaded by the
Buildkite nightly pipeline.