create a 'passphraseScheme' column when it's missing
Unfortunately, the manual migration are applied before the automated
migration. Yet, the passphraseScheme migration requires the column to
already exists in order to populate it. Therefore, we make sure to first
create it, and then, can safely assign it a value. If it already exists,
we simply assign a value.