use smaller delta for transaction list test.
One second is actually pretty large and we ended up in cases where,
adding one second to the time of the first transaction would make it
later than the second transaction! Causing some tests to fail for a
wrong reason. This delta needs to be at most the slot length, so I've
made it 100ms.