do not use implementation internal functions in property test for tx max size
The property was written by using the same function used for implementing the function it was testing. This is a recipe for errors for
the property will have the same bugs as the underlying function. Instead, I've rewritten properties to use only the outter function and
make statement about what can be observed. Also added some hard-coded classic unit tests to control the order of magnitude for the number
of inputs.