use slot + pool id as primary key for pool registration
This should prevent re-registration to blow up when being stored in the database. There can't possibly be
a pool with two 'valid' certificates on the same slot (the last one would prevail). I had to remove the f
foreign key on the metadata table because it was kinda awkward to link metadata to the pool id in the
first place, but even more awkward to also link the slot.