use 'Chan' instead of 'TQueue' for node's tip, and simplify the watcher logic
Chans have the nice property to be 'duplicable'. So, we can rely on this instead of re-implementing our own fan-out logic. I believe the original error that made worker crashes could have been a race-condition where the watcher would exit unexpectedly but I wasn't quite clearly able to identify the cause. Using 'Chan' looks definitely simpler.

I've already removed the possibility for workers to exit with a failure. This wasn't actually used and added on top of the complexity. If that's needed, we can work it out a bit later but for now, a simpler API makes it easier to debug.

Integration scenarios are no longer 'blowing up' unexpectedly, yet, the 'reward accumulate and stop' scenario still does not pass.